feat: Initialize user color via function
This commit updates the environment configuration to set the `$env.USER_COLOR` variable by calling the `init init_user_color` function.
Previously, the color was a hardcoded value. Moving this logic into a function improves modularity and centralizes initialization steps.
A minor adjustment was also made to the assignment of `$env.TRUE_HOME`.